Crate pallet_assets_precompiles

Re-exports

pub use foreign_assets::pallet;

pub use foreign_assets::pallet::Config as ForeignAssetsConfig;

pub use foreign_assets::ForeignAssetId;

pub use migration::MigrateForeignAssetPrecompileMappings;

pub use permit::pallet::Config as PermitConfig;

Modules

foreign_assets

migration

Migrations for pallet-assets-precompiles.

permit

ERC20Permit pallet for signature-based approvals (EIP-2612).

weights

Autogenerated weights for pallet_assets_precompiles

Structs

ERC20

An ERC20 precompile with EIP-2612 permit support.

ForeignAssetIdExtractor

An AssetIdExtractor that maps a local asset id (4 bytes taken from the address) to a foreign asset id.

ForeignIdConfig

A precompile configuration that uses a prefix AddressMatcher.

InlineAssetIdExtractor

An AssetIdExtractor that stores the asset id directly inside the address.

InlineIdConfig

A precompile configuration that uses a prefix AddressMatcher.

Traits

AssetIdExtractor

Means of extracting the asset id from the precompile address.

AssetPrecompileConfig

The configuration of a pallet-assets precompile.